[DESIGN] 상세정보 UI 완료
