+++ app/build.gradle
@@ -36,7 +36,7 @@ android {
 }
 
 dependencies {
-
+    implementation 'de.hdodenhof:circleimageview:3.1.0'
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.5.0'
