+++ app/src/main/res/values/colors.xml
@@ -4,4 +4,5 @@
     <color name="white">#FFFFFFFF</color>
     <color name="default_color">#00FFFFFF</color>
     <color name="light_grey">#C3C3C3</color>
+    <color name="grey">#D3D3D3</color>
 </resources>
